backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.124.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.20.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.2
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.236
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.38.0

# Additional dependencies for YOLO + DeepSeek-VL2 integration
accelerate>=0.26.0
sentencepiece>=0.1.99
einops>=0.7.0
timm>=0.9.12
